# openapi_final.json
{"openapi":"3.1.0","info":{"title":"Invoice Extractor API","version":"0.1.0"},"paths":{"/analyze-invoice":{"post":{"summary":"Analyze Invoice","description":"Step 1: Analyzes the invoice (OCR + Normalization) but DOES NOT persist to DB.","operationId":"analyze_invoice_analyze_invoice_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_invoice_analyze_invoice_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Analyze Invoice Analyze Invoice Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/confirm-invoice":{"post":{"summary":"Confirm Invoice","description":"Step 2: Persists validated data to Neo4j.","operationId":"confirm_invoice_confirm_invoice_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmInvoiceRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Confirm Invoice Confirm Invoice Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/report/{invoice_no}":{"get":{"summary":"Get Report","operationId":"get_report_report__invoice_no__get","parameters":[{"name":"invoice_no","in":"path","required":true,"schema":{"type":"string","title":"Invoice No"}}],"responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/invoices/{invoice_no}/items":{"get":{"summary":"Get Invoice Items Json","description":"Returns invoice details and line items as JSON (for Frontend Modal).","operationId":"get_invoice_items_json_invoices__invoice_no__items_get","parameters":[{"name":"invoice_no","in":"path","required":true,"schema":{"type":"string","title":"Invoice No"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/activity-log":{"get":{"summary":"Get Activity Log","description":"Returns a flat list of recent invoice activity for the Timeline View.","operationId":"get_activity_log_activity_log_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Get Activity Log Activity Log Get"}}}}}}},"/history":{"get":{"summary":"Get History Endpoint","description":"Returns grouped invoice history by Supplier.","operationId":"get_history_endpoint_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Get History Endpoint History Get"}}}}}}},"/enrich-supplier":{"post":{"summary":"Enrich Supplier Endpoint","description":"Triggers an Agent to find missing GST/Phone for a Supplier.","operationId":"enrich_supplier_endpoint_enrich_supplier_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EnrichmentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/inventory":{"get":{"summary":"Get Inventory Endpoint","description":"Returns aggregated inventory (Product Name + MRP).","operationId":"get_inventory_endpoint_inventory_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Response Get Inventory Endpoint Inventory Get"}}}}}}},"/export-excel":{"post":{"summary":"Export Excel Endpoint","description":"Generates an Excel report for the current data.","operationId":"export_excel_endpoint_export_excel_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_analyze_invoice_analyze_invoice_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_analyze_invoice_analyze_invoice_post"},"ConfirmInvoiceRequest":{"properties":{"invoice_data":{"additionalProperties":true,"type":"object","title":"Invoice Data"},"normalized_items":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Normalized Items"},"image_path":{"type":"string","title":"Image Path"}},"type":"object","required":["invoice_data","normalized_items"],"title":"ConfirmInvoiceRequest"},"EnrichmentRequest":{"properties":{"supplier_name":{"type":"string","title":"Supplier Name"}},"type":"object","required":["supplier_name"],"title":"EnrichmentRequest"},"ExportRequest":{"properties":{"invoice_data":{"additionalProperties":true,"type":"object","title":"Invoice Data"},"normalized_items":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Normalized Items"}},"type":"object","required":["invoice_data","normalized_items"],"title":"ExportRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}